+++ franka_example_controllers/include/franka_example_controllers/cartesian_orientation_example_controller.hpp
@@ -14,13 +14,12 @@
 
 #pragma once
 
-#include <Eigen/Dense>
 #include <string>
 
+#include <Eigen/Dense>
 #include <controller_interface/controller_interface.hpp>
 #include <rclcpp/rclcpp.hpp>
 
-#include <franka_example_controllers/default_robot_behavior_utils.hpp>
 #include <franka_semantic_components/franka_cartesian_pose_interface.hpp>
 
 using CallbackReturn = rclcpp_lifecycle::node_interfaces::LifecycleNodeInterface::CallbackReturn;
@@ -45,7 +44,7 @@ class CartesianOrientationExampleController : public controller_interface::Contr
 
  private:
   std::unique_ptr<franka_semantic_components::FrankaCartesianPoseInterface> franka_cartesian_pose_;
-  DefaultRobotBehavior default_robot_behavior_;
+
   Eigen::Quaterniond orientation_;
   Eigen::Vector3d position_;
   double trajectory_period_{0.001};