+++ franka_example_controllers/include/franka_example_controllers/cartesian_velocity_example_controller.hpp
@@ -19,7 +19,6 @@
 #include <controller_interface/controller_interface.hpp>
 #include <rclcpp/rclcpp.hpp>
 
-#include <franka_example_controllers/default_robot_behavior_utils.hpp>
 #include <franka_semantic_components/franka_cartesian_velocity_interface.hpp>
 
 using CallbackReturn = rclcpp_lifecycle::node_interfaces::LifecycleNodeInterface::CallbackReturn;
@@ -45,7 +44,6 @@ class CartesianVelocityExampleController : public controller_interface::Controll
  private:
   std::unique_ptr<franka_semantic_components::FrankaCartesianVelocityInterface>
       franka_cartesian_velocity_;
-  DefaultRobotBehavior default_robot_behavior_;
 
   const double k_time_max_{4.0};
   const double k_v_max_{0.05};