+++ franka_example_controllers/src/joint_velocity_example_controller.cpp
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#include <franka_example_controllers/default_robot_behavior_utils.hpp>
 #include <franka_example_controllers/joint_velocity_example_controller.hpp>
-#include <franka_msgs/srv/set_full_collision_behavior.hpp>
 
 #include <cassert>
 #include <cmath>
@@ -87,7 +87,7 @@ CallbackReturn JointVelocityExampleController::on_configure(
 
   auto client = get_node()->create_client<franka_msgs::srv::SetFullCollisionBehavior>(
       "service_server/set_full_collision_behavior");
-  auto request = default_robot_behavior_.getDefaultCollisionBehaviorRequest();
+  auto request = DefaultRobotBehavior::getDefaultCollisionBehaviorRequest();
 
   auto future_result = client->async_send_request(request);
   future_result.wait_for(1000ms);